PL0.Interpreter

Description

The PL/0 interpreter is a state machine that executes steps one by one and updates its own internal memory accordingly.

Scopes

data Scope

The Scope type represents the current state of the program.

Constructors

Scope

Fields

tables :: [SymbolTable]

symbols currentl visible

statements :: [Statement Identifier]

list of statements to execute

errors :: [String]

any errors that have been reported

prettyPrintScope :: Scope -> IO ()

Format a scope nicely and print it to stdout.

Symbols

type IdMap a = Map Identifier a

A polymorphic type for objects visible in the current scope.

data SymbolTable

A representation of all the objects visible in the current scope

Constructors

SymbolTable

Fields

constants :: IdMap Int

the constants currently visible

variables :: IdMap (Maybe Int)

the variables (initialized or non-initialized) currently visible

functions :: IdMap (Block Identifier)

the procedures currently visible

The interpreter

type Interpreter a = StateT Scope IO a

A state machine representing the process of executing a PL/0 program.

hasError :: Scope -> Bool

Check if any errors have been reported yet.

Running the interpreter

runProgram :: Program Identifier -> IO Scope

Run all the steps of a program. runProgram returns an IO action since we may need to get input from the user and print output to the screen. The IO action returns the final state of the interpreter.