PL0.Parser

Description

After the lexing stage, we need to transform a list of Tokens into an abstract syntax tree (AST) representing a full PL/0 program.

Parser type

type PL0Parser a = Parsec [Token] () a

A parser that represents a transformation from a list of Tokens.

If p is a parser of type PL0Parser (Block Identifier) and tks is a token list of type [Token], then parse p fp tks will give a value of type Either ParseError Block.

Instances of PL0Parser

parseExpression :: PL0Parser (Expression Identifier)

Parse an Expression.

parseTerm :: PL0Parser (Term Identifier)

Parse a Term.

parseFactor :: PL0Parser (Factor Identifier)

Parse a Factor.

parseOp :: PL0Parser Op

Parse a binary comparison operator.

parseCondition :: PL0Parser (Condition Identifier)

Parse a Condition.

parseStatements :: PL0Parser [Statement Identifier]

Parse a list of Statements.

parseConsts :: PL0Parser [(Identifier, Number)]

Parse a list of constant declarations.

parseVars :: PL0Parser [Identifier]

Parse a list of variable declarations.

parseProcs :: PL0Parser [(Identifier, Block Identifier)]

Parse a list of procedure definitions.

parseBlock :: PL0Parser (Block Identifier)

Parse a Block.

parseProgram :: PL0Parser (Program Identifier)

Parse a Program.