PL0.Syntax

Description

PL/0 follows a typical expression-factor-term model. For the full syntax specification in extended Backus-Naur form see here.

Structure of a PL/0 program

newtype Program a

Program is just a wrapper around Block.

Constructors

Program (Block a)

data Block a

The main body of a PL/0 program is a Block. It contains:

Zero or more constant declarations;

const x = 5;

Zero or more non-initialized variable declarations;

var square;

Zero or more procedure definitions;

procedure square;
begin
    squ:= x * x
end;

Exactly one Statement which is like the "main" function.

begin
    call square;
    ! squ;
end.

Constructors

Block

Fields

constDecls :: [(a, Number)]

Constant declarations

varDecls :: [a]

Variable declarations

procDefs :: [(a, Block a)]

Procedure declarations

body :: Statement a

Body statement

prettyPrintBlock :: Show a => Block a -> IO ()

Return a nice representation of a block, similar to the original PL/0 program.

Statements in PL/0

data Statement a

A statement in the PL/0 language.

Constructors

Set a (Expression a)
Set a variable to a value.

Call a
Call a procedure.

Input a
Read a number from stdin and store in a variable.

Output (Expression a)
Evaluate an expression and print the result to stdout.

StmtBlock [Statement a]
Execute a list of statements in series.

IfStmt (Condition a) (Statement a)
If the condition evaluates to true, execute the statement.

WhileStmt (Condition a) (Statement a)
While the condition evaluates to true, execute the statement.

PopBlock
Internal command used to return from procedure call.

prettyPrintStatement :: Show a => Int -> Statement a -> IO ()

Return a nice representation of a statement, similar to the original PL/0 program.

Expressions, terms and factors

data Expression a

An expression in the PL/0 language.

Constructors

UnaryPlus (Term a)
e.g. +t

UnaryMinus (Term a)
e.g. -t

BinaryPlus (Expression a) (Term a)
e.g. e+t

BinaryMinus (Expression a) (Term a)
e.g. e-t

data Term a

A term in the PL/0 language.

Constructors

SingleFactor (Factor a)
e.g. f

Mul (Term a) (Factor a)
e.g. t*f

Div (Term a) (Factor a)
e.g. t/f

data Factor a

A factor in the PL/0 language.

Constructors

Ident a
e.g. x

Num Number
e.g. 2

Parens (Expression a)
e.g. (e)

Conditions

data Condition a

An expression that evaluates to a boolean.

Constructors

Odd (Expression a)
Odd e evaluates to true if e evaluates to an odd number, else false.

Comp (Expression a) Op (Expression a)
Comp e1 op e2 checks if e1, e2 satisfy the condition represented by an Op type.

data Op

Enum describing binary comparison operators on two numbers.

Constructors

OP_LT

OP_LTE

OP_GT

OP_GTE

OP_EQ

OP_HASH

Type aliases

type Identifier = String

PL/0 strings are represented by the String datatype.

type Number = Int

PL/0 numbers are represented by the Int datatype.