PL0.Token

Description

PL0.Token exports a datatype Token which represents tokens used in the tokenizing stage.

Tokens used in parsing

data Token

Token is an intermediate datatype used in the tokenizing/lexing stage. There are different tokens representing keywords, punctuation, mathematical operators, and atomic values.

Constructors

Const

Var

Procedure

Call

Begin

End

If

Then

While

Do

Odd

Comma

Semicolon

Equals

ColonEquals

Dot

Question

Exclamation

Hash

LessThan

LessThanEquals

GreaterThan

GreaterThanEquals

Plus

Minus

Times

Divide

OpenParen

CloseParen

Identifier String

Number Int

Utility functions

sameType :: Token -> Token -> Bool

Check if two tokens are of the same type.