PL0

Description

PL/0 is an educational programming language. This package provides a library of datatypes and functions for tokenizing, lexing, parsing, interpreting, and compiling (WIP) PL/0 programs.

Tokenizing and lexing

module PL0.Token

module PL0.Lexer

Syntax and parsing

module PL0.Syntax

module PL0.Parser

Running a PL/0 program

module PL0.Interpreter