PL0.Lexer

Description

The lexing stage of the interpreter/compiler workflow transforms a String into an array of Tokens.

Lexing

lexPL0 :: Parser [Token]

A Parser representing a transformation of a String into a series of Tokens.